requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.3
pydantic~=2.8.2
PyYAML~=6.0.1
aiomysql~=0.2.0